requirements.txt
fastapi
uvicorn[standard]
python-chess
pydantic
networkx
numpy
scipy